Reorganize tree traits
The operational behaviour of a tree is defined by what was TreeWalk before.
Hence, TreeWalk has been renamed to Tree. The former Tree trait was removed on
the other hand, as the constructors it defines do not hold generally. One may
for example write a tree without associated data (such as the current Node) and
still implement Tree<P, N, O, ()> for it.