Move to associated types at large
The trait infrastructure now avoids parametrization over type parameters,
favouring associated types instead. To reuse associated type specifications,
ObjectQuery inherits Node and DataQuery inherits AssociatedData. Also note that
the Tree/PureTree traits have been removed, as they are fairly superfluous
given this new trait hierarchy.

This precludes some hypothetical use cases such as having a single tree
structure that can be viewed trivially as a tree in multiple senses. As it is
possible to create newtypes for such applications, the ergonomic benefits of
not having to deal with an explosion of type parameters outweigh this.

Fixes #16.