Use partition abstraction throughout the library
The tree implementations have been stripped completely of their specialized
geometry code. The trait infrastructure now also refers to partitions instead
of a node center and width.

Consequentially the util module was removed, as its functionality was no longer
relevant. Note though, that as a replacement new functions are required to
compute a partition that encompasses a set of positions.

The test suite is currently does not work, as the relevant partitioning schemes
have to be reimplemented.